+++ gregssnorkelscores/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 from .models import Location, Spot, Review, UserProfile
 
 class SpotAdmin(admin.ModelAdmin):
+    prepopulated_fields = {'slug':('name',)}
     list_display = ("name", "location", "url")
 
 class LocationAdmin(admin.ModelAdmin):
