+++ templates/gregssnorkelscores/add_location.html
@@ -6,4 +6,18 @@
 
 {% block body_block %}
     <h1>THIS IS WHERE THE FORM FOR ADD LOCATION WILL GO!</h1>
+    <div>
+        <form id="location_form" method="post" action="/gregssnorkelscores/add_location/">
+            {% csrf_token %}
+            {% for hidden in form.hidden_fields %}
+                {{ hidden }}
+            {% endfor %}
+            {% for field in form.visible_fields %}
+                {{ field.errors }}
+                {{ field.help_text }}
+                {{ field }}
+            {% endfor %}
+            <input type="submit" name="submit" value="Create Location" />
+        </form>
+    </div>
 {% endblock %}
