+++ templates/gregssnorkelscores/add_spot.html
@@ -6,4 +6,18 @@
 
 {% block body_block %}
     <h1>THIS IS WHERE THE FORM FOR ADD SNORKEL SPOT WILL GO!</h1>
+    <div> 
+        <form id="spot_form" method="post" action="/gregssnorkelscores/location_name/add_spot/">
+            {% csrf_token %}
+            {% for hidden in form.hidden_fields %}
+                {{ hidden }}
+            {% endfor %}
+            {% for field in form.visible_fields %}
+                {{ field.errors }}
+                {{ field.help_text }}
+                {{ field }}
+            {% endfor %}
+            <input type="submit" name="submit" value="Create Spot" />
+        </form>
+    </div>
 {% endblock %}
